+++ src/it/java/org/greencodeinitiative/creedengo/python/integration/tests/GCIRulesIT.java
@@ -273,6 +273,36 @@ class GCIRulesIT extends GCIRulesBase {
 
     }
 
+    @Test
+    void testGCI96() {
+        String filePath = "src/pandasRequireUsecols.py";
+        String ruleId = "creedengo-python:GCI96";
+        String ruleMsg = "Specify 'usecols' or 'columns' when reading a DataFrame using Pandas to load only necessary columns";
+        int[] startLines = new int[]{
+            3, 4, 5, 6, 7, 16, 19
+        };
+        int[] endLines = new int[]{
+            3, 4, 5, 6, 7, 16, 19
+        };
+
+        checkIssuesForFile(filePath, ruleId, ruleMsg, startLines, endLines, SEVERITY, TYPE, EFFORT_10MIN);
+    }
+
+    @Test
+    void testGCI97(){
+        String filePath = "src/optimizeSquareComputation.py";
+        String ruleId = "creedengo-python:GCI97";
+        String ruleMsg = "Use x*x instead of x**2 or math.pow(x,2) to calculate the square of a value";
+        int[] startLines = new int[]{
+            4, 7, 19, 20, 25, 26, 31, 38
+        };
+        int[] endLines = new int[]{
+            4, 7, 19, 20, 25, 26, 31, 38
+        };
+
+        checkIssuesForFile(filePath, ruleId, ruleMsg, startLines, endLines, SEVERITY, TYPE, EFFORT_1MIN);
+    }
+    
     @Test 
     void testGCI100() {
 
@@ -289,4 +319,36 @@ class GCIRulesIT extends GCIRulesBase {
         checkIssuesForFile(filePath, ruleId, ruleMsg, startLines, endLines, SEVERITY, TYPE, EFFORT_1h);
 
     }
+  
+    @Test
+    void testGCI103(){
+
+        String filePath = "src/dictionaryItemsUnused.py";
+        String ruleId = "creedengo-python:GCI103";
+        String ruleMsg = "Use dict.keys() or dict.values() instead of dict.items() when only one part of the key-value pair is used";
+        int[] startLines = new int[]{
+            5, 8, 24, 27, 36
+        };
+        int[] endLines = new int[]{
+            5, 8, 24, 27, 36
+        };
+
+        checkIssuesForFile(filePath, ruleId, ruleMsg, startLines, endLines, SEVERITY, TYPE, EFFORT_1MIN);  
+
+    }
+
+    @Test
+    void testGCI106() {
+        String filePath = "src/avoidSqrtInLoop.py";
+        String ruleId = "creedengo-python:GCI106";
+        String ruleMsg = "Avoid using scalar sqrt functions in loops. Apply vectorized sqrt operations on arrays directly.";
+        int[] startLines = new int[]{
+            7, 11, 16, 21, 45, 52, 60
+        };
+        int[] endLines = new int[]{
+            7, 11, 16, 21, 45, 52, 60
+        };
+        checkIssuesForFile(filePath, ruleId, ruleMsg, startLines, endLines, SEVERITY, TYPE, EFFORT_5MIN);
+    }
+
 }